station/requirements.txt
black==20.8b1
coverage==5.3.1
flask==1.1.2
imagezmq == 1.1.1
iniconfig==1.1.1
nose==1.3.7
numpy==1.19.5
opencv-contrib-python==4.5.1.48
numpy==1.19.5
pyzmq==22.0.2
py==1.10.0
pylint==2.6.0
